Refactor backend-specific converter branches
- move CoreML and llama-specific model lifecycle helpers out of KanaKanjiConverter.swift so the main converter flow reads as backend-agnostic logic
- switch ZenzCoreMLService from NSLock-based state management to an actor to make the CoreML path easier to reason about during review
- keep behavior unchanged while making the backend split more explicit for follow-up cleanup around shared Zenz APIs
