+++ Sources/KanaKanjiConverterModule/ConverterAPI/KanaKanjiConverter.swift
@@ -7,7 +7,6 @@
 //
 
 import Algorithms
-import Dispatch
 import EfficientNGram
 public import Foundation
 import SwiftUtils
@@ -48,65 +47,16 @@ public final class KanaKanjiConverter {
     public private(set) var zenzStatus: String = ""
     var dicdataStoreState: DicdataStoreState
 #if Zenzai
-    private var zenzaiModel: Zenz?
+    var zenzaiModel: Zenz?
 #endif
 #if ZenzaiCoreML && canImport(CoreML)
-    private var coreMLServiceStorage: Any?
-    private var zenzaiCoreMLCache: Kana2Kanji.ZenzaiCache?
-
-    private final class BlockingAsyncResultBox<T: Sendable>: @unchecked Sendable {
-        private let lock = NSLock()
-        private var value: T?
-
-        func store(_ value: T) {
-            self.lock.lock()
-            self.value = value
-            self.lock.unlock()
-        }
-
-        func load() -> T? {
-            self.lock.lock()
-            defer { self.lock.unlock() }
-            return self.value
-        }
-    }
-
-    @available(iOS 18, macOS 15, *)
-    private func resolvedCoreMLService() -> ZenzCoreMLService {
-        if let service = self.coreMLServiceStorage as? ZenzCoreMLService {
-            return service
-        }
-        let service = ZenzCoreMLService(owner: self)
-        self.coreMLServiceStorage = service
-        return service
-    }
-
-    @available(iOS 18, macOS 15, *)
-    private func blockingAsync<T: Sendable>(_ operation: @Sendable @escaping () async -> T) -> T {
-        let semaphore = DispatchSemaphore(value: 0)
-        let resultBox = BlockingAsyncResultBox<T>()
-        Task.detached {
-            resultBox.store(await operation())
-            semaphore.signal()
-        }
-        semaphore.wait()
-        return resultBox.load()!
-    }
+    var coreMLServiceStorage: Any?
+    var zenzaiCoreMLCache: Kana2Kanji.ZenzaiCache?
 #endif
 
     /// リセットする関数
     public func stopComposition() {
-#if ZenzaiCoreML && canImport(CoreML)
-        if #available(iOS 18, macOS 15, *), let service = self.coreMLServiceStorage as? ZenzCoreMLService {
-            self.blockingAsync {
-                await service.stopComposition()
-            }
-            self.coreMLServiceStorage = nil
-        }
-        self.zenzaiCoreMLCache = nil
-#elseif Zenzai
-        self.zenzaiModel = nil
-#endif
+        self.stopZenzBackendComposition()
         self.zenzaiPersonalization = nil
         self.previousInputData = nil
         self.lattice = .init()
@@ -131,38 +81,6 @@ public final class KanaKanjiConverter {
     func updateZenzStatus(_ text: String) {
         self.zenzStatus = text
     }
-
-#if ZenzaiCoreML && canImport(CoreML)
-    @available(iOS 18, macOS 15, *)
-    package func getModel(modelURL: URL) -> Zenz? {
-        self.blockingAsync {
-            await self.resolvedCoreMLService().getOrLoadModel(modelURL: modelURL)
-        }
-    }
-#elseif Zenzai
-    package func getModel(modelURL: URL) -> Zenz? {
-        if let cached = self.zenzaiModel, cached.resourceURL == modelURL {
-            self.updateZenzStatus("load \(modelURL.absoluteString)")
-            return cached
-        }
-        let model = self.blockingAsync {
-            try? await Zenz(resourceURL: modelURL)
-        }
-        if let model {
-            self.updateZenzStatus("load \(modelURL.absoluteString)")
-            self.zenzaiModel = model
-            return model
-        } else {
-            self.updateZenzStatus("zenz model unavailable")
-            return nil
-        }
-    }
-#else
-    package func getModel(modelURL: URL) -> Zenz? {
-        self.zenzStatus = "zenz-v2 model unavailable on this platform"
-        return nil
-    }
-#endif
 
     package func predictNextInputText(
         leftSideContext: String,
@@ -247,43 +165,6 @@ public final class KanaKanjiConverter {
             )
         }
     }
-
-#if ZenzaiCoreML && canImport(CoreML)
-    public func predictNextCharacter(leftSideContext: String, count: Int, options: ConvertRequestOptions) -> [(character: Character, value: Float)] {
-        guard #available(iOS 18, macOS 15, *) else {
-            print("zenz-v2 model unavailable")
-            return []
-        }
-        return self.blockingAsync {
-            await self.resolvedCoreMLService().predictNextCharacters(leftSideContext: leftSideContext, count: count, options: options)
-        }
-    }
-#elseif Zenzai
-    public func predictNextCharacterAsync(leftSideContext: String, count: Int, options: ConvertRequestOptions) async -> [(character: Character, value: Float)] {
-        guard options.zenzaiMode.versionDependentMode.version == .v2 else {
-            debug("next character prediction requires zenz-v2 models, not zenz-v1 nor zenz-v3 and later")
-            return []
-        }
-        guard let zenz = await self.getModel(modelURL: options.zenzaiMode.weightURL) else {
-            debug("zenz-v2 model unavailable")
-            return []
-        }
-        return await zenz.predictNextCharacter(leftSideContext: leftSideContext, count: count)
-    }
-
-    @available(*, deprecated, message: "Use async version 'predictNextCharacterAsync' instead to avoid blocking the calling thread")
-    nonisolated public func predictNextCharacter(leftSideContext: String, count: Int, options: ConvertRequestOptions) -> [(character: Character, value: Float)] {
-        let converter = self
-        return self.blockingAsync {
-            await converter.predictNextCharacterAsync(leftSideContext: leftSideContext, count: count, options: options)
-        }
-    }
-#else
-    public func predictNextCharacter(leftSideContext: String, count: Int, options: ConvertRequestOptions) -> [(character: Character, value: Float)] {
-        print("zenz-v2 model unavailable")
-        return []
-    }
-#endif
 
     /// 入力する言語が分かったらこの関数をなるべく早い段階で呼ぶことで、SpellCheckerの初期化が行われ、変換がスムーズになる
     public func setKeyboardLanguage(_ language: KeyboardLanguage) {
